requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
gunicorn>=21.2.0
datasets>=2.14.0